models

+++ diploma/src/main/java/iitu/kz/diploma/Models/Disciplines.java
@@ -15,19 +15,19 @@ public class Disciplines {
     @Column(name = "discipline", nullable = false, columnDefinition = "TEXT", unique = true)
     private String discipline;
 
-    @Column(name = "lecture_hours", nullable = false, columnDefinition = "BIGINT")
+    @Column(name = "lecture_hours", nullable = false, columnDefinition = "NUMERIC")
     private long lecture_hours;
 
-    @Column(name = "practice_hours", nullable = false, columnDefinition = "BIGINT")
+    @Column(name = "practice_hours", nullable = false, columnDefinition = "NUMERIC")
     private long practice_hours;
 
-    @Column(name = "laboratory_hours", nullable = false, columnDefinition = "BIGINT")
+    @Column(name = "laboratory_hours", nullable = false, columnDefinition = "NUMERIC")
     private long laboratory_hours;
 
-    @Column(name = "SIS", nullable = false, columnDefinition = "BIGINT")
+    @Column(name = "SIS", nullable = false, columnDefinition = "NUMERIC")
     private long sis;
 
-    @Column(name = "credits", nullable = false, columnDefinition = "BIGINT")
+    @Column(name = "credits", nullable = false, columnDefinition = "NUMERIC")
     private long credits;
 
     public Disciplines(){   }
