+++ diploma/src/main/java/iitu/kz/diploma/Repos/ArticlesRepo.java
@@ -3,5 +3,5 @@ package iitu.kz.diploma.Repos;
 import iitu.kz.diploma.Models.Articles;
 import org.springframework.data.repository.CrudRepository;
 
-public interface ArticlesRepo extends CrudRepository<Articles, Integer>{
+public interface ArticlesRepo extends CrudRepository<Articles, Long>{
 }
