+++ diploma/src/main/java/iitu/kz/diploma/Repos/DepartmentsRepo.java
@@ -3,5 +3,5 @@ package iitu.kz.diploma.Repos;
 import iitu.kz.diploma.Models.Departments;
 import org.springframework.data.repository.CrudRepository;
 
-public interface DepartmentsRepo extends CrudRepository<Departments, Integer> {
+public interface DepartmentsRepo extends CrudRepository<Departments, Long> {
 }
