+++ diploma/src/main/java/iitu/kz/diploma/Repos/DisciplinesRepo.java
@@ -3,6 +3,6 @@ package iitu.kz.diploma.Repos;
 import iitu.kz.diploma.Models.Disciplines;
 import org.springframework.data.repository.CrudRepository;
 
-public interface DisciplinesRepo extends CrudRepository<Disciplines, Integer>{
+public interface DisciplinesRepo extends CrudRepository<Disciplines, Long>{
 }
