+++ diploma/src/main/java/iitu/kz/diploma/Repos/FacultiesRepo.java
@@ -3,5 +3,5 @@ package iitu.kz.diploma.Repos;
 import iitu.kz.diploma.Models.Faculties;
 import org.springframework.data.repository.CrudRepository;
 
-public interface FacultiesRepo extends CrudRepository<Faculties, Integer>{
+public interface FacultiesRepo extends CrudRepository<Faculties, Long>{
 }
