+++ diploma/src/main/java/iitu/kz/diploma/Repos/GroupsRepo.java
@@ -3,5 +3,5 @@ package iitu.kz.diploma.Repos;
 import iitu.kz.diploma.Models.Groups;
 import org.springframework.data.repository.CrudRepository;
 
-public interface GroupsRepo extends CrudRepository<Groups, Integer> {
+public interface GroupsRepo extends CrudRepository<Groups, Long> {
 }
