+++ diploma/src/main/java/iitu/kz/diploma/Repos/SpecialitiesRepo.java
@@ -3,5 +3,5 @@ package iitu.kz.diploma.Repos;
 import iitu.kz.diploma.Models.Specialities;
 import org.springframework.data.repository.CrudRepository;
 
-public interface SpecialitiesRepo extends CrudRepository<Specialities, Integer> {
+public interface SpecialitiesRepo extends CrudRepository<Specialities, Long> {
 }
