+++ diploma/src/main/java/iitu/kz/diploma/Repos/TeachersRepo.java
@@ -3,5 +3,5 @@ package iitu.kz.diploma.Repos;
 import iitu.kz.diploma.Models.Teachers;
 import org.springframework.data.repository.CrudRepository;
 
-public interface TeachersRepo extends CrudRepository<Teachers, Integer> {
+public interface TeachersRepo extends CrudRepository<Teachers, Long> {
 }
