+++ diploma/src/main/java/iitu/kz/diploma/Repos/UsersRepo.java
@@ -3,5 +3,5 @@ package iitu.kz.diploma.Repos;
 import iitu.kz.diploma.Models.Users;
 import org.springframework.data.repository.CrudRepository;
 
-public interface UsersRepo extends CrudRepository<Users, Integer>{
+public interface UsersRepo extends CrudRepository<Users, Long>{
 }
